requirements.txt
apptools==4.5.0
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
envisage==4.9.0
httplib2==0.9.2
idna==2.6
imageio==2.6.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.1.2
mayavi==4.7.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
networkx==2.4
notebook==6.0.1
numpy==1.17.4
oauth==1.0.1
olefile==0.45.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.0.0
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyface==6.1.2
Pygments==2.5.2
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.5
pyRFC3339==1.0
pyrsistent==0.15.4
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.0.0
python-xlib==0.20
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
qtconsole==4.5.5
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.16.2
scipy==1.3.3
scour==0.36
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.13.0
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
traits==5.2.0
traitsui==6.1.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.7
vtk==8.1.2
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
